Adds integration tests
