+++ tests/fixtures/attic/gen3-bad.tsv
@@ -1,0 +1,3 @@
+ga4gh_drs_uri
+drs://FOOBAR:XXXX
+drs://FOOBAR:YYYY
