+++ tests/fixtures/attic/one-large-file.tsv
@@ -1,0 +1,2 @@
+entity:submitted_aligned_reads_id	pfb:created_datetime	pfb:data_category	pfb:data_format	pfb:data_type	pfb:experimental_strategy	pfb:file_format	pfb:file_md5sum	pfb:file_name	pfb:file_size	pfb:file_state	pfb:file_type	pfb:ga4gh_drs_uri	pfb:md5sum	pfb:object_id	pfb:project_id	pfb:read_group	pfb:state	pfb:submitter_id	pfb:updated_datetime
+01209272-c62f-4fdb-8cc1-89dc7f8e3f0e	2020-10-07T14:01:58.886297+00:00	Sequencing Data	CRAM	Aligned Reads	WGS	CRAM	d7ce3e6a5230ba4dd0d67740f34f90de	NA19648.final.cram	15896313948	registered	Aligned Reads	drs://dg.4503:dg.4503/21f3a4b9-d864-4ae3-a87c-e2daa689bdc5	d7ce3e6a5230ba4dd0d67740f34f90de	drs://dg.4503/21f3a4b9-d864-4ae3-a87c-e2daa689bdc5	open_access-1000Genomes	7a6540e7-7b57-4216-bf67-aef0b0be9e2f	validated	NA19648_cram	2021-05-05T20:58:36.989339+00:00
