+++ tests/fixtures/attic/one-small-file.tsv
@@ -1,0 +1,2 @@
+entity:aligned_reads_index_id	pfb:created_datetime	pfb:data_category	pfb:data_format	pfb:data_type	pfb:file_format	pfb:file_md5sum	pfb:file_name	pfb:file_size	pfb:file_state	pfb:file_type	pfb:ga4gh_drs_uri	pfb:md5sum	pfb:object_id	pfb:project_id	pfb:state	pfb:submitted_aligned_reads	pfb:submitter_id	pfb:updated_datetime
+00379748-7add-4e53-84ff-e0e5a92e4d72	2020-10-07T14:06:01.052463+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	b2b6648d0b6e9f04508afda4fb815e3e	HG02450.final.cram.crai	1405458	registered	Aligned Reads Index	drs://dg.4503:dg.4503/15fdd543-9875-4edf-8bc2-22985473dab6	b2b6648d0b6e9f04508afda4fb815e3e	drs://dg.4503/15fdd543-9875-4edf-8bc2-22985473dab6	open_access-1000Genomes	validated	38e0699f-84ef-4c50-a9fa-dde810a51d8e	HG02450_crai	2021-05-05T21:04:11.600109+00:00
