+++ tests/fixtures/attic/smol_big.tsv
@@ -1,0 +1,9 @@
+entity:submitted_aligned_reads_id	pfb:created_datetime	pfb:data_category	pfb:data_format	pfb:data_type	pfb:experimental_strategy	pfb:file_format	pfb:file_md5sum	pfb:file_name	pfb:file_size	pfb:file_state	pfb:file_type	pfb:ga4gh_drs_uri	pfb:md5sum	pfb:object_id	pfb:project_id	pfb:read_group	pfb:state	pfb:submitter_id	pfb:updated_datetime		
+0002ce34-0a06-426b-9104-4a0578cfaf8a	2020-10-07T14:02:48.604390+00:00	Sequencing Data	CRAM	Aligned Reads	WGS	CRAM	1b3ad8a0bab634829a11adfc965dd0df	HG03197.final.cram	15045663134	registered	Aligned Reads	drs://dg.4503:dg.4503/c64fded3-459b-4fca-82bc-35d154e9aa91	1b3ad8a0bab634829a11adfc965dd0df	drs://dg.4503/c64fded3-459b-4fca-82bc-35d154e9aa91	open_access-1000Genomes	9aa2160c-5bd7-44f2-ac12-b6bb8b695fd8	validated	HG03197_cram	2021-05-05T20:58:36.989339+00:00		
+01209272-c62f-4fdb-8cc1-89dc7f8e3f0e	2020-10-07T14:01:58.886297+00:00	Sequencing Data	CRAM	Aligned Reads	WGS	CRAM	d7ce3e6a5230ba4dd0d67740f34f90de	NA19648.final.cram	15896313948	registered	Aligned Reads	drs://dg.4503:dg.4503/21f3a4b9-d864-4ae3-a87c-e2daa689bdc5	d7ce3e6a5230ba4dd0d67740f34f90de	drs://dg.4503/21f3a4b9-d864-4ae3-a87c-e2daa689bdc5	open_access-1000Genomes	7a6540e7-7b57-4216-bf67-aef0b0be9e2f	validated	NA19648_cram	2021-05-05T20:58:36.989339+00:00		
+012114dc-fb94-490a-9fc2-589621e74bbc	2020-10-07T14:01:44.709753+00:00	Sequencing Data	CRAM	Aligned Reads	WGS	CRAM	9a37ae6c61f6ab194e1c7c48581e0408	NA18971.final.cram	16717892220	registered	Aligned Reads	drs://dg.4503:dg.4503/178bcdc0-3bbb-474c-bb28-fe3ace7f025b	9a37ae6c61f6ab194e1c7c48581e0408	drs://dg.4503/178bcdc0-3bbb-474c-bb28-fe3ace7f025b	open_access-1000Genomes	fc78a456-c69f-4fce-8d24-61d984efc009	validated	NA18971_cram	2021-05-05T20:58:36.989339+00:00		
+0777b34c-526c-43d9-b875-07dc968a6e32	2020-10-07T13:59:39.263741+00:00	Sequencing Data	CRAM	Aligned Reads	WGS	CRAM	e7e40194ed6f2fc0c36020439f41139c	HG01046.final.cram	15544000000	registered	Aligned Reads	drs://dg.4503:dg.4503/7ce2cdf2-47df-466b-b32d-70e951057251	e7e40194ed6f2fc0c36020439f41139c	drs://dg.4503/7ce2cdf2-47df-466b-b32d-70e951057251	open_access-1000Genomes	f3ad376d-56a8-46b8-993b-7fb12f542bda	validated	HG01046_cram	2021-05-05T20:58:42.944286+00:00		
+078d8c70-45b4-4d7a-b0ea-38e5a520b037	2020-10-07T14:01:02.822507+00:00	Sequencing Data	CRAM	Aligned Reads	WGS	CRAM	cdd0075aba09eea85bb6208b2b6ce87b	HG03809.final.cram	17410000000	registered	Aligned Reads	drs://dg.4503:dg.4503/41d96800-80f4-46a6-a38f-26dbaa36f0ac	cdd0075aba09eea85bb6208b2b6ce87b	drs://dg.4503/41d96800-80f4-46a6-a38f-26dbaa36f0ac	open_access-1000Genomes	efda0e7e-025d-40a0-91eb-7911e05d214c	validated	HG03809_cram	2021-05-05T20:58:42.944286+00:00		
+0795c810-a950-4b85-9995-c7793991b959	2020-10-07T14:02:34.708786+00:00	Sequencing Data	CRAM	Aligned Reads	WGS	CRAM	ea6e8dd8245f417b9ed18471e7ee36c6	HG02303.final.cram	15485000000	registered	Aligned Reads	drs://dg.4503:dg.4503/b6d93fc8-0e3b-44ad-bc6b-a70222176f0a	ea6e8dd8245f417b9ed18471e7ee36c6	drs://dg.4503/b6d93fc8-0e3b-44ad-bc6b-a70222176f0a	open_access-1000Genomes	c41e9b0f-2100-4e95-ae63-aabcdd15c481	validated	HG02303_cram	2021-05-05T20:58:42.944286+00:00		
+07a6e5a1-bd05-41e6-bc7a-c7d180b632e1	2020-10-07T14:01:16.503882+00:00	Sequencing Data	CRAM	Aligned Reads	WGS	CRAM	f0575f916848e4df6dc4f5becd97df69	HG04047.final.cram	19275000000	registered	Aligned Reads	drs://dg.4503:dg.4503/52b54dc0-b1c4-40b0-bc37-36e04f517654	f0575f916848e4df6dc4f5becd97df69	drs://dg.4503/52b54dc0-b1c4-40b0-bc37-36e04f517654	open_access-1000Genomes	0c5322f0-f3c6-40f9-9159-157625be04c1	validated	HG04047_cram	2021-05-05T20:58:42.944286+00:00		
+07c54dca-32fc-41f5-bc48-ff9a60c86918	2020-10-07T14:02:20.590992+00:00	Sequencing Data	CRAM	Aligned Reads	WGS	CRAM	03f8625a1c176099a1e235c642ac5647	NA21115.final.cram	19031000000	registered	Aligned Reads	drs://dg.4503:dg.4503/2c0d11d7-3ffd-4840-8797-1bcfe7a058c7	03f8625a1c176099a1e235c642ac5647	drs://dg.4503/2c0d11d7-3ffd-4840-8797-1bcfe7a058c7	open_access-1000Genomes	48d53dac-202f-4ad1-994b-4183e8126640	validated	NA21115_cram	2021-05-05T20:58:42.944286+00:00
