+++ tests/fixtures/gen3-small.tsv
@@ -1,0 +1,17 @@
+ga4gh_drs_uri
+drs://FOOBAR:d62af3ff-756d-4e73-b6c6-e0bbd3bae50d
+drs://FOOBAR:755fb48b-c4c9-41b6-a3fd-2604ebaa21af
+drs://FOOBAR:cdff48c3-69a6-4c24-84d8-82681803003f
+drs://FOOBAR:da8cb14b-8274-4b5a-a5ed-279140c9cf74
+drs://FOOBAR:82808bc8-a21b-43bc-af61-fcbc0437cd13
+drs://FOOBAR:82b3224d-98a5-453c-ab50-edb96e72fda6
+drs://FOOBAR:93f7ca81-93dc-4046-a537-40a36465062c
+drs://FOOBAR:f4b29275-499f-4260-8725-d6c9f5d5bcb8
+drs://FOOBAR:839de941-4317-4508-968e-83883590bec9
+drs://FOOBAR:80164516-695a-473f-9c76-eacb9e191a6a
+drs://FOOBAR:c5196828-493b-4f62-976a-46468088547e
+drs://FOOBAR:6d5df2b1-5811-4206-a53b-7eac9e8f6383
+drs://FOOBAR:f905ecc0-e865-47fc-9d62-9087f5a4150e
+drs://FOOBAR:9ed5fcf8-c461-46ad-84a3-12d709d378d4
+drs://FOOBAR:892f99d6-8cdf-432b-bc8b-7d965d0cad53
+drs://FOOBAR:34819aba-29f7-4a83-b5d4-84b10a3f6b1b
