+++ tests/fixtures/terra-data-bad.tsv
@@ -1,0 +1,11 @@
+entity:aligned_reads_index_id	pfb:created_datetime	pfb:data_category	pfb:data_format	pfb:data_type	pfb:file_format	pfb:file_md5sum	pfb:file_name	pfb:file_size	pfb:file_state	pfb:file_type	pfb:ga4gh_drs_uri	pfb:md5sum	pfb:object_id	pfb:project_id	pfb:state	pfb:submitted_aligned_reads	pfb:submitter_id	pfb:updated_datetime
+00379748-7add-4e53-84ff-e0e5a92e4d72	2020-10-07T14:06:01.052463+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	b2b6648d0b6e9f04508afda4fb815e3e	HG02450.final.cram.crai	1405458	registered	Aligned Reads Index	drs://dg.BAD_DRS:dg.BAD_DRS/15fdd543-9875-4edf-8bc2-22985473dab6	b2b6648d0b6e9f04508afda4fb815e3e	drs://dg.BAD_DRS/15fdd543-9875-4edf-8bc2-22985473dab6	open_access-1000Genomes	validated	38e0699f-84ef-4c50-a9fa-dde810a51d8e	HG02450_crai	2021-05-05T21:04:11.600109+00:00
+003b7e95-90e9-4c84-a2d7-96f334b403fc	2020-10-07T14:06:43.903478+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	155e057a069e438ed5b091d2bdfbfe58	HG03873.final.cram.crai	1351946	registered	Aligned Reads Index	drs://dg.BAD_DRS:dg.BAD_DRS/3c861ec6-d810-4058-b851-c0b19dd5933e	155e057a069e438ed5b091d2bdfbfe58	drs://dg.BAD_DRS/3c861ec6-d810-4058-b851-c0b19dd5933e	open_access-1000Genomes	validated	36e24428-5c59-481d-910f-ae060192e22d	HG03873_crai	2021-05-05T21:04:11.600109+00:00
+0049197b-a446-4d0f-b8ca-20c56544e42f	2020-10-07T14:05:10.971075+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	750004f80de56fb9e74bde5a252b0260	HG00536.final.cram.crai	1244278	registered	Aligned Reads Index	drs://dg.BAD_DRS:dg.BAD_DRS/374a0ad9-b3a2-47f3-8860-5083b302e478	750004f80de56fb9e74bde5a252b0260	drs://dg.BAD_DRS/374a0ad9-b3a2-47f3-8860-5083b302e478	open_access-1000Genomes	validated	7cbfb95a-361f-4419-8094-ef260535fa52	HG00536_crai	2021-05-05T21:04:11.600109+00:00
+00824a20-a84d-4c29-b24e-b537dad9d934	2020-10-07T14:06:59.420707+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	2d9f53cf93b214c8b9eb396b78cd41aa	NA18613.final.cram.crai	1370106	registered	Aligned Reads Index	drs://dg.BAD_DRS:dg.BAD_DRS/231527ef-7075-477f-9265-4a1289576384	2d9f53cf93b214c8b9eb396b78cd41aa	drs://dg.BAD_DRS/231527ef-7075-477f-9265-4a1289576384	open_access-1000Genomes	validated	4ad88f65-409f-4130-aaf6-ec4c47ffca1d	NA18613_crai	2021-05-05T21:04:11.600109+00:00
+00b3f719-55fd-4be5-a0c4-8e27a6ebb68b	2020-10-07T14:05:54.052719+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	499c5bdcdb0c8f72a49954fac62208a1	HG02142.final.cram.crai	1405543	registered	Aligned Reads Index	drs://dg.BAD_DRS:dg.BAD_DRS/6f64fe3d-f62a-46ca-b1f8-77b3754db409	499c5bdcdb0c8f72a49954fac62208a1	drs://dg.BAD_DRS/6f64fe3d-f62a-46ca-b1f8-77b3754db409	open_access-1000Genomes	validated	d22183ac-2269-4988-92de-e7f59fb0fc23	HG02142_crai	2021-05-05T21:04:11.600109+00:00
+00b898c2-1bb0-4d26-a389-7c370e8aaca4	2020-10-07T14:07:34.051570+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	1bfa7aef44300f86847e5d6c57cf4805	NA20356.final.cram.crai	1368064	registered	Aligned Reads Index	drs://dg.BAD_DRS:dg.BAD_DRS/6a4dfbfe-5fff-43a6-8824-ffe2990b9180	1bfa7aef44300f86847e5d6c57cf4805	drs://dg.BAD_DRS/6a4dfbfe-5fff-43a6-8824-ffe2990b9180	open_access-1000Genomes	validated	e5c2ba5b-bbfc-4f96-a18f-cc6d599af145	NA20356_crai	2021-05-05T21:04:11.600109+00:00
+00c38711-68bf-4e65-a7cf-201969e29a89	2020-10-07T14:05:17.833139+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	238eb9fce97703ae1b9b6b6aaa00b0f3	HG00622.final.cram.crai	1254920	registered	Aligned Reads Index	drs://dg.BAD_DRS:dg.BAD_DRS/1b387791-53fb-48fa-9594-d51953814f19	238eb9fce97703ae1b9b6b6aaa00b0f3	drs://dg.BAD_DRS/1b387791-53fb-48fa-9594-d51953814f19	open_access-1000Genomes	validated	cd822a58-75d0-4e19-9d16-d04cb8ebe1bb	HG00622_crai	2021-05-05T21:04:11.600109+00:00
+00cd257e-66f7-4614-89b3-cb60dff9f19d	2020-10-07T14:08:03.123969+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	a1c0992a39316b7a4cd928860bcacb21	HG01552.final.cram.crai	1296198	registered	Aligned Reads Index	drs://dg.BAD_DRS:dg.BAD_DRS/fdb50b7d-f126-4ddb-93f8-6c2d8faf25d3	a1c0992a39316b7a4cd928860bcacb21	drs://dg.BAD_DRS/fdb50b7d-f126-4ddb-93f8-6c2d8faf25d3	open_access-1000Genomes	validated	db17b8db-41dd-482d-9d9b-8db6e980c309	HG01552_crai	2021-05-05T21:04:11.600109+00:00
+00eb3fff-15a6-403e-918d-cc338d20d6d3	2020-10-07T14:07:40.584651+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	91c78edd69237db468723c797d77bd30	NA20525.final.cram.crai	1337382	registered	Aligned Reads Index	drs://dg.BAD_DRS:dg.BAD_DRS/4f1de21e-53cf-439e-a867-2179408c2187	91c78edd69237db468723c797d77bd30	drs://dg.BAD_DRS/4f1de21e-53cf-439e-a867-2179408c2187	open_access-1000Genomes	validated	72d514f4-b9bd-4746-939f-ba1824e16a01	NA20525_crai	2021-05-05T21:04:11.600109+00:00
+0117d1e8-42d9-43ed-bb03-057568443f4f	2020-10-07T14:06:57.997587+00:00	Sequencing Data	CRAI	Aligned Reads Index	CRAI	dddc01550f967f9097adae3c6252e785	HG04209.final.cram.crai	1338980	registered	Aligned Reads Index	drs://dg.BAD_DRS:dg.BAD_DRS/bf2b854a-17a3-4b3c-aeb2-4f670ceb9e85	dddc01550f967f9097adae3c6252e785	drs://dg.BAD_DRS/bf2b854a-17a3-4b3c-aeb2-4f670ceb9e85	open_access-1000Genomes	validated	1f91b3c1-7822-4cbc-82e7-2a3516088c10	HG04209_crai	2021-05-05T21:04:11.600109+00:00
